+++ packages/core/index.ts
@@ -32,6 +32,7 @@ export { inMemoryCache } from './src/cache/adapters/in_memory';
 export { localStorageCache } from './src/cache/adapters/local_storage';
 export { sessionStorageCache } from './src/cache/adapters/session_storage';
 export { voidCache } from './src/cache/adapters/void';
+export { createCacheAdapter } from './src/cache/adapters/instance';
 
 // Exposed libs
 export {