+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farfetched/solid
 
+## 0.8.6
+
+## 0.8.5
+
+## 0.8.4
+
 ## 0.8.3
 
 ### Patch Changes